requirements/base.txt
django==3.1.6
psycopg2-binary==2.8.6
djangorestframework==3.12.2
djangorestframework-gis==0.17
celery==4.4.7
django-redis==4.12.1

--extra-index-url https://gitlab+deploy-token-215547:tneDsrg6in6tG_wM8N9B@gitlab.com/api/v4/projects/24487626/packages/pypi/simple
service-core==0.0.13

Pillow==8.1.0
jsonschema==3.2.0
requests==2.25.1
